coin_converter/requirements.txt
pyyaml==6.0.3
requests==2.32.5
streamlit==1.51.0